requirements.txt
discord==1.0.1
mariadb==1.0.5
python-dotenv==0.15.0
Quart~=0.14.1